requirements.txt
requests>=2.31.0
PyYAML>=6.0
boto3>=1.26.0  # Optional: For S3-based version storage
